% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{ctl}
\alias{ctl}
\title{calculate LC/LC}
\usage{
ctl(df)
}
\arguments{
\item{df}{A dataframe with three variables(concentration/dose, death, total), the first observation is control.}
}
\description{
Used for control. If death rate in control group < 0.05, no correction is needed;
if death rate between 0.05 - 0.2, using Abbott fomula to adjust the assay;
if death rate > 0.2, invalid assay. Program will stop! Private function in LD50 package.
}
