% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{LD_cal}
\alias{LD_cal}
\title{main function to calculate LD50.}
\usage{
LD_cal(dfr)
}
\arguments{
\item{dfr}{A dataframe with three variables(concentration/dose, death, total), make sure the first observation(row) is control.}
}
\description{
If death rate in control group < 0.05, no correction is needed;
if death rate between 0.05 - 0.2, using Abbott fomula to adjust the assay;
if death rate > 0.2, invalid assay. Program will stop!
}
\examples{
aa <- data.frame('con' = c(0,0.01,0.02,0.04,0.08,0.16,0.32), 
'death' = c(1,6,16,23,25,34,44), 'total' = c(60,59,60,60,57,58,60))
LD_cal(aa)

}
