% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_fun.R
\name{LDx}
\alias{LDx}
\title{calulate LDx}
\usage{
LDx(fit.model, x)
}
\arguments{
\item{fit.model}{a object of glm model}

\item{x}{which death rate, eg, LD50 -> x = 0.5}
}
\description{
LDx is used to infer LD50 and corresponding CI.
For private use in package LD50.
}
